update: rename package
