+++ converter/openaiclaudeconverter/claude_to_openai.go
@@ -1,4 +1,4 @@
-package openaiclaude
+package openaiclaudeconverter
 
 import (
 	"strings"
