+++ converter/openaiclaudeconverter/openai_to_claude.go
@@ -1,4 +1,4 @@
-package openaiclaude
+package openaiclaudeconverter
 
 import (
 	"fmt"
